cryptohome: Update LeafIndex of LoggedRecord from int to int64_t
LeafIndex is actually defined as int64 from the server side.

BUG=b:232747549
TEST=cros_run_unit_tests --board=${BOARD} --packages cryptohome

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3892357
Commit-Queue: Sherri Lin <sherrilin@google.com>
Tested-by: Sherri Lin <sherrilin@google.com>

+++ cryptohome/cryptorecovery/recovery_crypto_util.h
@@ -57,7 +57,7 @@ struct LoggedRecord {
   // Serialized private log entry.
   brillo::Blob private_log_entry;
   // Leaf index of this record in the tree.
-  int leaf_index = -1;
+  int64_t leaf_index = -1;
 };
 
 // LedgerSignedProof is included in HsmResponseAssociatedData.
